open music api v2

+++ src/services/postgres/PlaylistsService.js
@@ -4,11 +4,12 @@ const InvariantError = require('../../exceptions/InvariantError')
 const NotFoundError = require('../../exceptions/NotFoundError')
 const AuthorizationError = require('../../exceptions/AuthorizationError')
 const { playlistsModel } = require('../../utils/model')
+const { playlistActivitiesModel } = require('../../utils/model')
 
 class PlaylistsService {
-  constructor () {
+  constructor (collaborationsService) {
     this._pool = new Pool()
-    // this._collaborationsService = collaborationsService
+    this._collaborationsService = collaborationsService
   }
 
   async addPlaylist ({ name, owner }) {
@@ -30,24 +31,19 @@ class PlaylistsService {
 
   async getPlaylists (owner) {
     const query = {
-      // text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
-      text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner WHERE playlists.owner = $1',
+      text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
       values: [owner]
     }
 
     const result = await this._pool.query(query)
 
-    if (!result.rowCount) {
-      throw new NotFoundError('Playlist tidak ditemukan')
-    }
-
     return result.rows
   }
 
-  async deletePlaylistById (id) {
+  async deletePlaylistById (id, owner) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
-      values: [id]
+      text: 'DELETE FROM playlists WHERE id = $1 AND owner = $2 RETURNING id',
+      values: [id, owner]
     }
 
     const result = await this._pool.query(query)
@@ -104,20 +100,39 @@ class PlaylistsService {
     return result.rows
   }
 
-  async deleteSongFromPlaylistById (id, songId) {
+  async deleteSongFromPlaylistById (playlistId, songId) {
     const query = {
-      text: 'DELETE FROM playlist_songs WHERE playlist_id = $1 AND song_id = $2',
-      values: [id, songId]
+      text: 'DELETE FROM playlist_songs WHERE playlist_id = $1 AND song_id = $2 RETURNING id',
+      values: [playlistId, songId]
     }
 
     const result = await this._pool.query(query)
 
     if (!result.rowCount) {
-      throw new NotFoundError('Lagu gagal dihapus dari album. Id tidak ditemukan')
+      throw new NotFoundError('Lagu gagal dihapus dari playlist. Id tidak ditemukan')
     }
   }
 
   async verifyPlaylistOwner (id, owner) {
+    const query = {
+      text: 'select playlists.id, playlists.name, playlists.owner, collaborations.user_id from playlists LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id WHERE playlists.id = $1',
+      values: [id]
+    }
+
+    const result = await this._pool.query(query)
+
+    if (!result.rowCount) {
+      throw new NotFoundError('Playlist tidak ditemukan')
+    }
+
+    const playlist = result.rows[0]
+
+    if (playlist.owner !== owner && playlist.user_id !== owner) {
+      throw new AuthorizationError('Anda tidak berhak mengakses resource ini')
+    }
+  }
+
+  async verifyPlaylistAccess (id, owner) {
     const query = {
       text: 'SELECT * FROM playlists WHERE id = $1',
       values: [id]
@@ -136,13 +151,9 @@ class PlaylistsService {
     }
   }
 
-  async verifyPlaylistAccess (playlistId, userId) {
-    await this._collaborationsService.verifyCollaborator(playlistId, userId)
-  }
-
   async isPlaylistExist (id) {
     const query = {
-      text: 'SELECT * FROM playlists WHERE id = $1',
+      text: 'SELECT id FROM playlists WHERE id = $1',
       values: [id]
     }
 
@@ -151,6 +162,8 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new NotFoundError('Playlist tidak ditemukan')
     }
+
+    return result.rows[0].id
   }
 
   async isSongExist (id) {
@@ -164,6 +177,37 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new NotFoundError('Lagu tidak ditemukan')
     }
+  }
+
+  async addPaylistHistory (playlistId, songId, userId, action) {
+    const id = `activities-${nanoid(16)}`
+    const time = new Date().toISOString()
+
+    const query = {
+      text: 'INSERT INTO playlist_song_activities VALUES ($1, $2, $3, $4, $5, $6) RETURNING id',
+      values: [id, playlistId, songId, userId, action, time]
+    }
+
+    const result = await this._pool.query(query)
+
+    if (!result.rowCount) {
+      throw new InvariantError('Activities gagal ditambahkan')
+    }
+  }
+
+  async getPlaylistActivities (playlistId) {
+    const query = {
+      text: 'SELECT psa.playlist_id, u.username, s.title, action, time FROM playlist_song_activities psa LEFT JOIN playlists p ON p.id = psa.playlist_id LEFT JOIN users u ON u.id = psa.user_id LEFT JOIN songs s ON s.id = psa.song_id WHERE psa.playlist_id = $1',
+      values: [playlistId]
+    }
+
+    const result = await this._pool.query(query)
+
+    if (!result.rowCount) {
+      throw new NotFoundError('Activity tidak ditemukan')
+    }
+
+    return result.rows.map(playlistActivitiesModel)
   }
 }
